@@ -1,16 +1,22 @@
 {
-  "name": "gherking-template",
+  "name": "gpc-template",
   "version": "0.0.1",
   "description": "This repository is a template to create precompilers for GherKing.",
-  "main": "lib/index.js",
+  "main": "index.ts",
+  "types": "index.d.ts",
   "scripts": {
-    "test": "mocha ./test/*.spec.js ./test/builtIn/*.spec.js",
-    "coverage": "istanbul cover --include-all-sources ./node_modules/mocha/bin/_mocha --root ./lib -- ./test/*.spec.js ./test/builtIn/*.spec.js",
-    "travis": "istanbul cover --include-all-sources ./node_modules/mocha/bin/_mocha --root ./lib --report lcovonly -- ./test/*.spec.js ./test/builtIn/*.spec.js -R spec && cat ./coverage/lcov.info | node ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage"
+    "build": "npm run buildUpdate && npm test && npm run typedoc",
+    "buildUpdate": "npm run clean && npm run compile && npm run copyToDist",
+    "copyToDist": "copyfiles -f *.txt *.md package.json dist",
+    "typedoc": "typedoc --out ./docs ./src && touch ./docs/.nojekyll",
+    "clean": "rimraf ./dist ./docs ./reports ./coverage",
+    "test": "cross-env JEST_JUNIT_OUTPUT_DIR=./reports jest --coverage",
+    "lint": "eslint . --ext .ts --fix",
+    "compile": "tsc && npm run lint"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/gherking/gherking-template.git"
+    "url": "git+https://github.com/gherking/gpc-template.git"
   },
   "keywords": [
     "gherkin",
@@ -22,21 +28,71 @@
   "author": "GherKing",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/gherking/gherking-template/issues"
+    "url": "https://github.com/gherking/gpc-template/issues"
   },
+  "homepage": "https://github.com/gherking/gpc-template#readme",
+  "files": [
+    "**/*.js",
+    "**/*.d.ts",
+    "**/*.js.map"
+  ],
   "engines": {
-    "node": ">=6.0.0",
-    "npm": ">=3.0.0"
+    "node": ">=12.0.0"
   },
   "devDependencies": {
-    "chai": "^4.1.2",
-    "coveralls": "^3.0.0",
-    "istanbul": "^0.4.5",
-    "mocha": "^5.0.0"
+    "@types/jest": "^26.0.20",
+    "@types/node": "^14.14.21",
+    "@typescript-eslint/eslint-plugin": "^4.14.1",
+    "@typescript-eslint/parser": "^4.14.1",
+    "copyfiles": "^2.4.1",
+    "cross-env": "^7.0.3",
+    "eslint": "^7.19.0",
+    "jest": "^26.6.3",
+    "jest-junit": "^12.0.0",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^26.4.4",
+    "typedoc": "^0.20.16",
+    "typescript": "^4.1.3"
   },
-  "homepage": "https://github.com/gherking/gherking-template#readme",
   "dependencies": {
-    "gherkin-ast": "^1.1.0",
-    "gherkin-precompiler": "^1.3.0"
+    "gherkin-ast": "^3.0.1",
+    "gherking": "file:../gherking/gherking-1.0.0.tgz"
+  },
+  "resolutions": {
+    "gherkin-ast": "3.0.1"
+  },
+  "jest": {
+    "reporters": [
+      "default",
+      "jest-junit"
+    ],
+    "transform": {
+      "^.+\\.ts?$": "ts-jest"
+    },
+    "testEnvironment": "node",
+    "testMatch": [
+      "**/tests/**/*.test.ts"
+    ],
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "dist"
+    ],
+    "collectCoverage": true,
+    "coverageDirectory": "./coverage",
+    "collectCoverageFrom": [
+      "./src/**/*.ts"
+    ],
+    "moduleFileExtensions": [
+      "ts",
+      "js"
+    ],
+    "coverageThreshold": {
+      "global": {
+        "branches": 95,
+        "functions": 85,
+        "lines": 85,
+        "statements": 85
+      }
+    }
   }
 }
