@@ -1,10 +1,13 @@
-const { DefaultConfig } = require('gherkin-precompiler');
-const { /* TODO */ } = require('gherkin-ast');
+import { PreCompiler } from "gherking";
+import { /* TODO */ } from "gherkin-ast";
 
 // TODO: Add implementation of your precompiler
+export class Template implements PreCompiler {
+
+}
 /*
  * @example:
- * class MyPrecompiler extends DefaultConfig {
+ * export class MyPrecompiler implements PreCompiler {
  *   constructor(config) {
  *     super();
  *     this.config = config;
